quark-d2000-devboard: Add support for watchdog
This patch implements the watchdog APIs from Contiki (core/dev/watchdog.h)
and enables the watchdog during platform initialization. The watchdog is
started right before user applications are started.

The watchdog is configured with timeout equals to 8.389 seconds. This
value is arbitrary and can be changed in future. So any process that
takes more than 8 seconds executing should call watchdog_periodic() in
order to avoid the system to reboot.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>