quark-d2000-devboard: build: Add 'debug' rule
This patch adds the '--debug' option to flasher script so we are
able to setup debugging sessions very easily. This option basically
runs 'openocd' and 'gdb' with the right parameters.

It also adds the 'debug' rule to our build system which uses this
new flasher option and extends the README file so it provides
instructions about starting debugging sessions.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>